Vercel.json - vercel server path clean
