+++ vercel.json
@@ -1,0 +1,5 @@
+{
+    "rewrites": [
+      { "source": "/(.*)", "destination": "/" }
+    ]
+}
